csvlite connector for csvparser
