+++ airbyte-integrations/connectors/source-csv-lite/src/test/java/CSVConnectorLiteTest.java
@@ -33,7 +33,7 @@ public class CSVConnectorLiteTest {
         String metricStoreURL =  "http://anom-metric-store.bha.svc.cluster.local:4242/api/anoms/api/put?details";
         String uniqueIdentifier = UUID.randomUUID().toString();
         String token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJST0xFIjoiQVBJIiwic3ViIjoic3VtaXQtdGVzdCIsIk9SR19JRCI6IjY0IiwiaXNzIjoic3VtaXRAYmljeWNsZS5pbyIsImlhdCI6MTY0NDk0MTQxMywiVEVOQU5UIjoiZW10LWU5ZTRlZjZjLTYzYzQtNDkzMC1iMzMxLTJkZjNhZjFlNzg4ZSIsImp0aSI6IjBkZjU4ZmFkLTk0NzMtNDQ4OS1iNzMifQ.t8F2oEwEFej1xU2LknY2pLsbgUW3x5YED8trN9QYzDU";
-        String connectorId = "73b631a0-3c28-403a-bfaf-44e93e4c385f";
+        String connectorId = "d804c589-d03e-4429-873c-2ee006db495b";
         String userId = "";
         String eventSourceType= "EVENT";
         String tenantId = "emt-e9e4ef6c-63c4-4930-b331-2df3af1e788e";
@@ -81,13 +81,13 @@ public class CSVConnectorLiteTest {
         csvConnector.setBicycleEventProcessorAndPublisher(bicycleConfig);
     }
 
-    //@Test
+    @Test
     public void testSyncData() throws Exception {
         csvConnector.syncData(config, catalog, new ObjectMapper().createObjectNode(), null);
         Assertions.assertTrue(true);
     }
 
-    //@Test
+    @Test
     public void testRead() throws Exception {
         csvConnector.doRead(config, catalog, new ObjectMapper().createObjectNode());
         Assertions.assertTrue(true);
